% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BanditPAM-package.R
\docType{package}
\name{banditpam-package}
\alias{banditpam-package}
\title{banditpam is a package for fast clustering using medoids}
\description{
banditpam is a high-performance package for almost
linear-time k-medoids clustering).
}
\author{
anon and anon
}
\keyword{package}
