% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bpam_num_threads}
\alias{bpam_num_threads}
\title{Return the number of threads banditpam is using}
\usage{
bpam_num_threads()
}
\value{
the number of threads banditpam is using
}
\description{
Return the number of threads banditpam is using
}
