% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.KMedoids__fit}
\alias{.KMedoids__fit}
\title{Fit the KMedoids algorthm given the data and loss}
\usage{
.KMedoids__fit(xp, data, loss, distMat = NULL)
}
\arguments{
\item{xp}{the km::KMedoids Object XPtr}

\item{data}{the data matrix}

\item{loss}{the loss indicator}

\item{distMat}{the optional distance matrix}
}
\description{
Fit the KMedoids algorthm given the data and loss
}
