% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.KMedoids__get_k}
\alias{.KMedoids__get_k}
\title{Return the number of medoids property k}
\usage{
.KMedoids__get_k(xp)
}
\arguments{
\item{xp}{the km::KMedoids Object XPtr}
}
\description{
Return the number of medoids property k
}
