% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.KMedoids__get_loss_fn}
\alias{.KMedoids__get_loss_fn}
\title{Return the loss_fn property}
\usage{
.KMedoids__get_loss_fn(xp)
}
\arguments{
\item{xp}{the km::KMedoids Object XPtr}
}
\description{
Return the loss_fn property
}
