% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.KMedoids__get_max_iter}
\alias{.KMedoids__get_max_iter}
\title{Return the max_iter property}
\usage{
.KMedoids__get_max_iter(xp)
}
\arguments{
\item{xp}{the km::KMedoids Object XPtr}
}
\description{
Return the max_iter property
}
