% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.KMedoids__get_medoids_final}
\alias{.KMedoids__get_medoids_final}
\title{Return the final medoids}
\usage{
.KMedoids__get_medoids_final(xp)
}
\arguments{
\item{xp}{the km::KMedoids Object XPtr}
}
\description{
Return the final medoids
}
