% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.KMedoids__get_statistic}
\alias{.KMedoids__get_statistic}
\title{Return specified metric/statistics from the computation}
\usage{
.KMedoids__get_statistic(xp, what)
}
\arguments{
\item{xp}{the km::KMedoids Object XPtr}

\item{what}{which metric to return, 1 = "dist_computations", 2 = "dist_computations_and_misc", 3 = "misc_dist", 4 = "build_dist", 5 = "swap_dist", 6 = "cache_writes", 7 = "cache_hits", 8 = "cache_misses"}
}
\description{
Return specified metric/statistics from the computation
}
