% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.KMedoids__get_swap_conf}
\alias{.KMedoids__get_swap_conf}
\title{Return the swap_conf property}
\usage{
.KMedoids__get_swap_conf(xp)
}
\arguments{
\item{xp}{the km::KMedoids Object XPtr}
}
\description{
Return the swap_conf property
}
