% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.KMedoids__new}
\alias{.KMedoids__new}
\title{Create a new KMedoids object.}
\usage{
.KMedoids__new(k, max_iter, build_confidence, swap_confidence)
}
\value{
an external ptr (Rcpp::XPtr) to a KMedoids object instance.
}
\description{
Create a new KMedoids object.
}
