% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.KMedoids__set_iter}
\alias{.KMedoids__set_iter}
\title{Set the max_iter property}
\usage{
.KMedoids__set_iter(xp, bc)

.KMedoids__set_iter(xp, bc)

.KMedoids__set_iter(xp, bc)
}
\arguments{
\item{xp}{the km::KMedoids Object XPtr}
}
\description{
Set the max_iter property

Set the build_conf property

Set the swap_conf property
}
