% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.KMedoids__set_k}
\alias{.KMedoids__set_k}
\title{Set the number of medoids property k}
\usage{
.KMedoids__set_k(xp, k)
}
\arguments{
\item{xp}{the km::KMedoids Object XPtr}
}
\description{
Set the number of medoids property k
}
