% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.KMedoids__set_loss_fn}
\alias{.KMedoids__set_loss_fn}
\title{Set the loss_fn property}
\usage{
.KMedoids__set_loss_fn(xp, loss_fn)
}
\arguments{
\item{xp}{the km::KMedoids Object XPtr}
}
\description{
Set the loss_fn property
}
